Fix: Holding left and right at the same time is now visually correct
For the purpose of this example, I will use the Prev and Next keys,
though this fix also applies to the Feed and Skip keys.
When holding left, the Prev text moves to the left.  When holding
right, the Next text moves to the right.  However, if you hold Prev,
then hold Next (so you're holding both at the same time), the Prev
text snaps back and only the Next text is moved.  This fix allows
both texts to be moved at the same time.