Feature: Add method to return sprite surfaces
Before, the only way to draw a sprite was to use the draw() method.
Now, you can use the render() method instead.  This method doesn't
draw directly onto a surface.  Instead, it returns a new surface
which contains the sprite you want to draw.  The main advantage to
this method over the draw() method is that you can manipulate this
sprite surface as much as you want, before blitting it somewhere else.