Fix: Add properties to graphics that I forgot to add
When I made a lot of the graphics attributes private, I forgot to
actually test that the code works afterwards.  Needless to say,
my IDE did not detect some of the uses of certain attributes which
I made private, and so the code crashed on startup.  I have
now fixed that by adding the necessary properties.