Feature: Add ability to load a singular sprite
Before, you had to create a SpriteColumn of length 1 if you wanted
a singular sprite.  But since singular sprites are so common, as
well as the fact that single_width, single_height, and single_size
(the size of a single sprite) were redundant when there was only
one sprite, I decided to make a class for singular sprites.