Feature: Add png loading to files.py
Every image is currently loaded with its own string literal:
"images/file_name.png".  This is no good, because what if we change
the folder the images are in?  What if we decide .jpg is a better
file format?  We'd have to change every single string literal.  It
would be better if all the images were loaded from the same method.