Feature: Make save file compatible even with wrong number of levels
Previously, if any levels were added or removed, the number of levels
tracked in the save file would have been different to the amount of
levels in the game, eventually causing an error to be raised for one
reason or another.  This change makes it so that the game works even
when these two level counts are different, and (if the save_progress
method is called) modifies the save file so that it lines up with
the game.