Refactor: Change "lethal" attribute to "has_deadly_effect"
In the jam version, the only thing that made a bottle lethal for
certain was having a deadly side effect, so a lethal bottle directly
implied that the bottle had a deadly side effect.  However, now there
are more ways a bottle can be lethal, and so the title "lethal"
doesn't work anymore.  Thus the renaming.