Feature: Make bottle palettes store their own name
Instead of having a comment beside each palette that states its
name, the name is now stored as a string attribute in the Palette
object.

This is in preparation for moving the palettes into a JSON file.
Without a name, it would be hard to tell what each of the palettes
in the JSON file is, since they'll only be differentiated by their
RGB values (which look a lot like arbitrary numbers).