Feature: Remove the Lime Lime palette
I didn't really like this palette because of how similar it is
to Lime Grey, as well as the fact that there are already enough
bottles with the same color of grey as the body color.

Note that removing a palette shifts the index of every palette
after it down by one.  Since the menu bottles pick their
palette based on the index of the palette, I had to go and
decrease all those indexes by one.