Refactor: Move the deletion of excess result screen bottles
If you spam feed after you feed a bottle that makes you win/lose,
the game may add the extra bottles you fed to the results screen.
This is somewhat confusing since your last bottle will be
different from what you thought it was, so the extra bottles
are deleted.  This refactor moves the deletion of these bottles
into the _win() and _lose() methods themselves.  This allows
you to override the _win() and _lose() methods to remove the
extra bottle deletion if so desired.