Feature: Implement feed/skip lock
Previously, the feed/skip locking was only necessary when in an
ending cutscene, so I used the _in_ending_cutscene boolean to decide
whether or not to lock feed/skip.  However, in the future I may
require the locking to occur in other situations.  Thus, I have
created a new boolean called _feed_skip_locked which you can set to
True to lock or False to unlock.