Fix: Make bottle judgement no longer based on animation
Bottle judgement used to be done whenever the homunculus reached a
certain frame of its animation.  In a previous commit, that was replaced
by a queue of pending judgements, with timers that would just so happen
to align with the animation.  However, I accidentally kept in a
condition where this queue of judgements would only update while the
homunculus was in its eating animation.  Thus, if you managed to feed
bottles in a way that the eating animation ends before all bottles
are judged, you could end up leaving judgements frozen in the queue.

This bug requires you to feed bottles in quick succession, something
which almost always kills you in normal mode.  However, in race mode,
you cannot die, which means that this bug actually has a noticeable
effect.  In fact, after the race mode mistake showcase was implemented
this actually softlocked the game.  That is why it needed to be fixed
ASAP.