Feature: Make Curves store whether they are active or not
Previously, you had to use some indirect method to determine whether a
curve animation was active or not.  Either checking what frame of the
animation it is on, or storing an external boolean.  Now, I've simply
added a boolean to the Curve itself which you can use to store whether
it's enabled or not.