Fix: Make Curve animations inclusive on end of animation
Earlier, I wrote in the documentation that the starting value of a
curve animation occurs on frame 0 and that the ending value occurs
on the last frame.  However, that is not true: the ending value
actually occurs one frame after the last frame.  This disparity
between the documentation and the code is now fixed.

Why do I want it to be inclusive on both sides?  Personally, it just
makes sense for frame 0 to be the first value and the last frame to be
the last value.  Also, the animation stops being active once the last
frame is reached, after which asking for the value of the curve would
give you the value before the final value instead of the final value
itself.