Fix: Increase SHIFT_LENGTH to accommodate curve changes
Since Curve animations have been changed so that the last frame's value
is equal to the final value of the animation (previously, it was the
penultimate value of the animation instead), every Curve animation has
"shortened" by one frame.  Because of this, some of the animations
looked a little off.  To fix this, I have increased one of the constants
(which defines the length of some animations) by 1.