Fix: Add missing image "left_text.png"
This image exists in my local repository, but I forgot to add it.
Pulling the repository (without this image) and trying to run the game
would result in an error.  This image contains the characters in the
"Bottles Left!" text in race mode, and was supposed to be added in
one of the commits implementing race mode.