Refactor: Clear up ambiguity of bottle shifting methods
Previously, the names of these methods were "_shift_left()" and
"_shift_right()".  This is ambiguous, since (in the case of _shift_left)
it could either mean the bottles move to the left, or that the focus
shifts to the bottle on the left (which means the bottles move to the
right).  Thus, I have renamed them to "_shift_to_next_bottle()" and
"_shift_to_previous_bottle()" to clear up this ambiguity.