Refactor: Make BottleScreen bottles list a property
This is so that the getters and setters of the bottle list can be
overwritten by child classes.  This is useful because BottleIconRow
needs a reference to BottleScreen's bottle list, but if the bottle
list is set to something else, BottleIconRow's reference won't update
alongside with it.  This can be fixed if you overwrite BottleScreen's
bottle list setter so that it updates BottleIconRow's reference.