Fix: Prevent crashing when cycling through allergens
This has to do with the list of allergies in the results screen.
Sometimes, when scrolling backwards through the results, the game would
crash trying to remove a nonexistent allergy from the list.  This was
caused by an error in making RaceScreen extend SelectionScreen.
Previously, RaceScreen would change current_bottle before trying to
remove all of current_bottle's allergies.  However, in the new
implementation, current_bottle doesn't change, and so the allergies that
were attempted to be removed were from the wrong bottle.  This has been
fixed by selecting the bottle preceding current_bottle in the list, and
removing that bottle's allergies instead of current_bottle's.