Fix: Remove allergies added by final bottle from allergy list
The allergies list on ResultScreen adds the allergies of the current
bottle only when you advance to the next bottle.  That means that the
allergies added by the very last bottle should never show up in the
list, since there is no next bottle after it.  However, the way
allergies are transferred from PlayScreen to ResultScreen adds the
allergy of the last bottle, and since the game never expects this
allergy to be added, it never attempts to remove it either, meaning it's
permanently stuck on the list.

This commit fixes that by explicitly removing the allergies added by the
final bottle from the list.