Fix: Make icon row's selected bottle change before updating scroll
This fixes a crash that would sometimes occur after losing a stage.
_icon_row's selected bottle number should always match BottleScreen's
current_bottle_num, however, that is not always the case.  Since the
code updates the scrolling of _icon_row before it makes these two
variables match, for one frame _icon_row will scroll to the previous
value of current_bottle_num before fixing itself.  This means that
if the previous value of current_bottle_num is invalid (higher than the
last index of the bottles list), the game will crash.

This commit fixes this by reordering the scroll update and the variable
matching so that the variable matching always comes first.